Move to io.opentelemetry.semconv:opentelemetry-semconv (#1044)
