+++ maven-extension/src/main/java/io/opentelemetry/maven/handler/SnykTestHandler.java
@@ -8,7 +8,7 @@ package io.opentelemetry.maven.handler;
 import io.opentelemetry.api.trace.SpanBuilder;
 import io.opentelemetry.api.trace.SpanKind;
 import io.opentelemetry.maven.MavenGoal;
-import io.opentelemetry.semconv.trace.attributes.SemanticAttributes;
+import io.opentelemetry.semconv.SemanticAttributes;
 import java.util.Collections;
 import java.util.List;
 import org.apache.maven.execution.ExecutionEvent;
@@ -25,9 +25,9 @@ final class SnykTestHandler implements MojoGoalExecutionHandler {
   public void enrichSpan(SpanBuilder spanBuilder, ExecutionEvent executionEvent) {
     spanBuilder.setSpanKind(SpanKind.CLIENT);
     spanBuilder.setAttribute(SemanticAttributes.PEER_SERVICE, "snyk.io");
-    spanBuilder.setAttribute(SemanticAttributes.HTTP_URL, "https://snyk.io/api/v1/test-dep-graph");
+    spanBuilder.setAttribute(SemanticAttributes.URL_FULL, "https://snyk.io/api/v1/test-dep-graph");
     spanBuilder.setAttribute(SemanticAttributes.RPC_METHOD, "test");
-    spanBuilder.setAttribute(SemanticAttributes.HTTP_METHOD, "POST");
+    spanBuilder.setAttribute(SemanticAttributes.HTTP_REQUEST_METHOD, "POST");
   }
 
   @Override
